rename packages correctly
