+++ test/e2e/logs/multi_pipeline_max_pipeline_test.go
@@ -1,6 +1,6 @@
 //go:build e2e
 
-package fluentbit
+package logs
 
 import (
 	"fmt"
