+++ test/e2e/logs/otel/application/multi_pipeline_namespace_selector_test.go
@@ -1,6 +1,6 @@
 //go:build e2e
 
-package otel
+package application
 
 import (
 	. "github.com/onsi/ginkgo/v2"
