+++ test/e2e/logs/otel/application/suite_test.go
@@ -1,6 +1,6 @@
 //go:build e2e
 
-package otel
+package application
 
 import (
 	"testing"
