+++ test/e2e/logs/otel/otlp/basic_test.go
@@ -1,6 +1,6 @@
 //go:build e2e
 
-package otel
+package otlp
 
 import (
 	"io"
